Split magic wand to its own component
